feat: moved RichText and UserMention out of Intepretation components
RichText has the Jest tests from the old d2-ui implementation.
