+++ src/components/Interpretations/InterpretationModal/CommentAddForm.js
@@ -3,8 +3,8 @@ import i18n from '@dhis2/d2-i18n'
 import { Button } from '@dhis2/ui'
 import PropTypes from 'prop-types'
 import React, { useRef, useState } from 'react'
+import { RichTextEditor } from '../../RichText/index.js'
 import {
-    RichTextEditor,
     MessageEditorContainer,
     MessageButtonStrip,
     MessageInput,
