--- src/components/Interpretations/common/RichTextEditor/index.js
@@ -1,1 +1,0 @@
-export { RichTextEditor } from './RichTextEditor.js'
