+++ src/components/Interpretations/common/index.js
@@ -1,4 +1,3 @@
 export * from './Interpretation/index.js'
 export * from './Message/index.js'
-export * from './RichTextEditor/index.js'
 export * from './getInterpretationAccess.js'
