+++ src/components/RichText/Editor/markdownHandler.js
