+++ src/components/RichText/Editor/styles/Editor.style.js
